+++ tests/testVec3.cxx
@@ -1,4 +1,4 @@
-#include <MvVector3.h>
+#include <mntVec3.h>
 
 int main() {
